requirements.txt
streamlit==1.37.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
pdfplumber==0.11.4
python-docx==1.1.2
pydantic==2.8.2
transformers==4.43.4
accelerate==0.34.2
huggingface-hub==0.24.5
numpy==1.26.4
scikit-learn==1.5.1
orjson==3.10.7
requests==2.32.3